Revert Game2 base flow and expose reel view
